adding ir support
